+++ packages/validate/src/hooks/use-form-validator-messages.tsx
@@ -2,9 +2,16 @@
 import { GetAllObjectPaths, GetPathObjectType } from '@under-control/core';
 import { ValidationError, ValidationErrorsListProps } from '../types';
 
+type FormExtractorAttrs = {
+  includeGlobals?: boolean;
+};
+
 type FormErrorsExtractor<V> = <P extends GetAllObjectPaths<V>>(
   path: P,
+  attrs?: FormExtractorAttrs,
 ) => ValidationErrorsListProps<GetPathObjectType<V, P>>;
+
+type GlobalFormErrorsExtractor<V> = () => ValidationErrorsListProps<V>;
 
 export type FormValidatorMessagesHookAttrs<V> = {
   errors: Array<ValidationError<V>>;
@@ -12,30 +19,39 @@ export type FormValidatorMessagesHookAttrs<V> = {
 
 export type FormValidatorMessagesHookResult<V> = {
   all: Array<ValidationError<V>>;
+  global: GlobalFormErrorsExtractor<V>;
   extract: FormErrorsExtractor<V>;
 };
 
 export function useFormValidatorMessages<V>({
   errors,
 }: FormValidatorMessagesHookAttrs<V>): FormValidatorMessagesHookResult<V> {
-  const extract: FormErrorsExtractor<V> = path => ({
+  const extract: FormErrorsExtractor<V> = (path, attrs) => ({
     errors: errors.flatMap(item => {
       const itemPath = item.path as string;
-      if (itemPath && itemPath !== path) {
+      if (
+        (!itemPath || itemPath !== path) &&
+        (!attrs?.includeGlobals || itemPath)
+      ) {
         return [];
       }
 
       return [
         {
           ...item,
-          path,
+          path: itemPath,
         } as unknown as ValidationError<any>,
       ];
     }),
   });
 
+  const global: GlobalFormErrorsExtractor<V> = () => ({
+    errors: errors.filter(error => !(error.path as string)),
+  });
+
   return {
     all: errors,
+    global,
     extract,
   };
 }