+++ packages/core/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @under-control/core
 
+## 1.2.1
+
+### Patch Changes
+
+- Fixed error extractor
+
 ## 1.2.0
 
 ### Minor Changes