+++ packages/forms/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @under-control/forms
 
+## 1.2.1
+
+### Patch Changes
+
+- Fixed error extractor
+- Updated dependencies
+  - @under-control/validate@1.2.1
+  - @under-control/core@1.2.1
+  - @under-control/inputs@1.2.1
+
 ## 1.2.0
 
 ### Minor Changes