+++ packages/inputs/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @under-control/inputs
 
+## 1.2.1
+
+### Patch Changes
+
+- Fixed error extractor
+- Updated dependencies
+  - @under-control/core@1.2.1
+
 ## 1.2.0
 
 ### Minor Changes